% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_info.R
\name{comment_info}
\alias{comment_info}
\title{Extracts information on comments including course, year, comments, number of thumbsups and number of thumbsdowns.}
\usage{
comment_info(url, y = 2018)
}
\arguments{
\item{url}{A Character value indicating the url of the webpage corresponding to an instructor.}

\item{y}{A number indicating the user are interested in comments after that year.}
}
\description{
Extracts information on comments including course, year, comments, number of thumbsups and number of thumbsdowns.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
comment_info(url = url, y = 2018)
}
