% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_schools.R
\name{get_all_schools}
\alias{get_all_schools}
\title{URL by School Name Generator}
\usage{
get_all_schools(school_name)
}
\arguments{
\item{school_name}{A character value indicating the name of the university.}
}
\value{
A character string as the URL corresponding to the university of interest.
}
\description{
Finds the university URL using its name.
}
\examples{
get_all_schools('Iowa State University')
}
