% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tid.R
\name{get_tid}
\alias{get_tid}
\title{Gets teacher IDs}
\usage{
get_tid(name, department = NULL, university = NULL)
}
\arguments{
\item{name}{A character value which is the name of an instructor}
}
\value{
A tibble of teacher ID and general information
}
\description{
Gets teacher IDs and general information by name.
}
\examples{
name <- "Brakor"
department <- "Biology"
university <- "California Berkeley"
get_tid(name = name)
get_tid(name = name, department = department)
get_tid(name = name, university = university)
get_tid(name = name, department = department, university = university)
}
