% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{URL Extractor}
\usage{
get_url(name, department = NULL, university = NULL)
}
\arguments{
\item{name}{A character value of the professor's full name (required).}

\item{department}{A character value of the professor's department (optional).}

\item{university}{A character value of the professor's university (required).}
}
\value{
A vector of URL(s)
}
\description{
Extracts URL by the name, department, and university of the professor.
}
\examples{
name <- "Brakor"
department <- "Biology"
university <- "California Berkeley"
get_url(name = name, university = university)
get_url(name = name, department = department, university = university)
}
