% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attendance.R
\name{plot_attendance}
\alias{plot_attendance}
\title{Plot a bar plot for the attendance type for an instructor}
\usage{
plot_attendance(url)
}
\arguments{
\item{url}{A character value corresponding to the instructor of interest}
}
\description{
Plot a bar plot for the attendance type for an instructor
}
\examples{
plot_attendance("https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2353616")
}
