% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QualDiff.R
\name{plot_QualDiff}
\alias{plot_QualDiff}
\title{Creates a bar plot for the quality and difficulty of a course for an instructor}
\usage{
plot_QualDiff(url)
}
\arguments{
\item{url}{A character value corresponding to the instructor of interest}
}
\description{
Creates a bar plot for the quality and difficulty of a course for an instructor
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
plot_QualDiff(url = url)

}
