% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_info.R
\name{ratings_info}
\alias{ratings_info}
\title{Ratings Summarizer}
\usage{
ratings_info(url, y = 2018)
}
\arguments{
\item{url}{A character value indicating the URL of the webpage corresponding to an instructor.}

\item{y}{A number indicating the user are interested in ratings after that year.}
}
\value{
A list of the number of ratings after filtering and 2 data frames
\itemize{
  \item n - Number of ratings after filtering
  \item ratings - All rating information after a given year
  \item summary - Summary statistics
}
}
\description{
Extracts and summarizes all rating information for an instructor.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
ratings_info(url = url)
}
