% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_plot.R
\name{ratings_plot}
\alias{ratings_plot}
\title{Ratings Plot}
\usage{
ratings_plot(url, y = 2018)
}
\arguments{
\item{url}{A character value indicating the url of the webpage corresponding to an instructor}

\item{y}{A number indicating the user are interested in ratings after that year.}
}
\value{
A plot including 4 subplots
}
\description{
A boxplot of all ratings and 3 barplots of average ratings by course, grade, and year.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
ratings_plot(url = url)
}
