% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_summ.R
\name{ratings_summ}
\alias{ratings_summ}
\title{Ratings Summarizer}
\usage{
ratings_summ(url)
}
\arguments{
\item{url}{A Character value indicating the url of the webpage corresponding to an instructor}
}
\value{
A list with eight elements
\itemize{
  \item n_ratings - Number of ratings
  \item percent_take_again - Percentage of students that would take again
  \item difficulty - Summary statistics for difficulty ratings
  \item take_again_logic - A logical vector indicating the response of taking again
  \item for_credit_logic - A logical vector indicating the response of for credit
  \item textbooks_logic - A logical vector indicating the response of textbooks needed
  \item attendance_logic - A logical vector indicating the response of attendance needed
  \item grade_key - A character vector showing students' grades
}
}
\description{
Summarize all rating information for an instructor
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
ratings_summ(url = url)
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=1801125'
ratings_summ(url = url)
}
