% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_info.R
\name{sentiment_info}
\alias{sentiment_info}
\title{Sentiment Info Extractor}
\usage{
sentiment_info(url, y = 2018, word = "Positive")
}
\arguments{
\item{url}{A character value indicating the URL of the professor's webpage.}

\item{y}{A numeric value to filter ratings after a certain year.}

\item{word}{A string indicating the user is interested in positive words, negative words, or tags. Choices among "Positive", "Negative", and "Tags". (default="positive")}
}
\value{
A data frame with 2 columns
\itemize{
  \item word - Words or tags
  \item n - Count of words in comments or tags
}
}
\description{
Provides common positive words and negative words extracted from comments, and tags from the website.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
sentiment_info(url = url, y = 2018, word = "Negative")
}
