% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_plot.R
\name{sentiment_plot}
\alias{sentiment_plot}
\title{Sentiment Plot}
\usage{
sentiment_plot(url, y = 2018, word = "Positive")
}
\arguments{
\item{url}{A character value indicating the URL of the professor's webpage.}

\item{y}{A numeric value to filter ratings after a certain year.}

\item{word}{A character value indicating the user's interest in positive words, negative words, or tags.}
}
\value{
A wordcloud of common words or tags
}
\description{
A word cloud of common positive words and negative words extracted from comments, and tags from the website.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
sentiment_plot(url = url, y = 2018, word = "Positive")
}
