% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_info.R
\name{comment_info}
\alias{comment_info}
\title{Comment Info Extractor}
\usage{
comment_info(url, y = 2018)
}
\arguments{
\item{url}{A character value indicating the URL of the webpage corresponding to an instructor.}

\item{y}{A number indicating the user is interested in comments after that year.}
}
\value{
A data frame with 5 columns
\itemize{
  \item course - Course code
  \item year - Delivery year for the course
  \item comments - Comments for the course
  \item thumbsup - Number of thumbs-up
  \item thumbsdown - Number of thumbs-down
}
}
\description{
Extracts information on comments including course, year, comments, number of thumbs-ups and number of thumbs-downs.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
comment_info(url = url, y = 2018)
}
