% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_info.R
\name{comment_info}
\alias{comment_info}
\title{Comments Extractor}
\usage{
comment_info(url, y = 2018)
}
\arguments{
\item{url}{A character value indicating the URL of the professor's webpage.}

\item{y}{A numeric value to filter comments after a certain year.}
}
\value{
A data frame with 5 columns
\itemize{
  \item course - Course code
  \item year - Delivery year of the course
  \item comments - Comments on the professor
  \item thumbsup - Number of thumbs-up
  \item thumbsdown - Number of thumbs-down
}
}
\description{
Extracts comments together with course, year, number of thumbs-ups, and number of thumbs-downs, and filters by year.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
comment_info(url = url, y = 2018)
}
