% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_info.R
\name{comment_info}
\alias{comment_info}
\title{Provides wordcloud plots for positive words, negative words in the comments, and tags on the website.}
\usage{
comment_info(url, y = 2018, word = "Positive")
}
\arguments{
\item{url}{A Character value indicating the url of the webpage corresponding to an instructor.}

\item{y}{A number indicating the user are interested in comments after that year.}

\item{word}{A string indicating the user are interested in positive words, or negative words, or tags.}
}
\description{
Provides wordcloud plots for positive words, negative words in the comments, and tags on the website.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
comment_info(url = url, y = 2018, word = "Negative")
}
