% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_info.R
\name{general_info}
\alias{general_info}
\title{General Info Finder}
\usage{
general_info(url)
}
\arguments{
\item{url}{A character value indicating the URL of the webpage corresponding to an instructor.}
}
\value{
A list with two elements
\itemize{
  \item name - Complete name of the instructor
  \item department - The department of the instructor
  \item university - The unversity of the instructor
}
}
\description{
Extracts general information for an instructor.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
general_info(url = url)
}
