% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_info.R
\name{general_info}
\alias{general_info}
\title{General Info Finder}
\usage{
general_info(url)
}
\arguments{
\item{url}{A Character value indicating the url of the webpage corresponding to an instructor}
}
\value{
A list with two elements
\itemize{
  \item name - Complete Name of the Instructor
  \item place - The department and university of the instructor
}
}
\description{
Extracts general information for an instructor
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
general_info(url = url)
}
