% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_url_school.R
\name{get_all_schools}
\alias{get_all_schools}
\title{Finds the university url using its name}
\usage{
get_all_schools(school_name)
}
\arguments{
\item{school_name}{A Character value indicating the name of the university}
}
\value{
A url corresponding to the university of interest
}
\description{
Finds the university url using its name
}
\examples{
get_all_schools('Iowa State University')
get_all_schools('University of Iowa')
}
