% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tid.R
\name{get_tid}
\alias{get_tid}
\title{Gets teacher IDs}
\usage{
get_tid(name)
}
\arguments{
\item{name}{A character value which is the name of an instructor}
}
\value{
A tibble of teacher ID and general information
}
\description{
Gets teacher IDs and general information by name.
}
\examples{
name <- "Brakor"
get_tid(name = name)
}
