% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tid.R
\name{get_tid}
\alias{get_tid}
\title{Get teacher ID}
\usage{
get_tid(name)
}
\arguments{
\item{name}{A character value which is the name of an instructor}
}
\value{
A tibble of teacher ID and general information
}
\description{
Get the teacher ID instructors by name
}
\examples{
name <- "A"
get_tid(name = name)
}
