% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tid.R
\name{get_tid}
\alias{get_tid}
\title{Professor ID Extractor}
\usage{
get_tid(name, department = NULL, university)
}
\arguments{
\item{name}{A character value of the professor's full name (required).}

\item{department}{A character value of the professor's department (optional).}

\item{university}{A character value of the professor's university (required).}
}
\value{
A tibble of the professor's ID and the general information with 4 columns
\itemize{
  \item tID - ID of the professor
  \item name - Full name of the professor
  \item department - The department of the professor
  \item university - The university of the professor
}
}
\description{
Extracts the professor's ID and combines it with general information.
}
\examples{
name <- "Brakor"
department <- "Biology"
university <- "California Berkeley"
get_tid(name = name, university = university)
get_tid(name = name, department = department, university = university)
}
