% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tid.R
\name{get_tid}
\alias{get_tid}
\title{Professor ID Extractor}
\usage{
get_tid(name, department = NULL, university = NULL)
}
\arguments{
\item{name}{A character value of the professor's name.}

\item{department}{A character value of the professor's department.}

\item{university}{A character value of the professor's university.}
}
\value{
A tibble of the professor's ID and the general information with 4 columns
\itemize{
  \item tID - ID of the professor
  \item name - Complete name of the professor
  \item department - The department of the professor
  \item university - The unversity of the professor
}
}
\description{
Extracts professor's ID and combines with general information.
}
\examples{
name <- "Brakor"
department <- "Biology"
university <- "California Berkeley"
get_tid(name = name)
get_tid(name = name, department = department)
get_tid(name = name, university = university)
get_tid(name = name, department = department, university = university)
}
