% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Gets url}
\usage{
get_url(name, department = NULL, university = NULL)
}
\arguments{
\item{name}{A character value which is the name of an instructor}

\item{department}{A character value which is the department of an instructor}

\item{university}{A character value which is the university of an instructor}
}
\value{
A vector of urls
}
\description{
Gets url by name, department and university.
}
\examples{
name <- "Brakor"
department <- "Biology"
university <- "California Berkeley"
get_url(name = name)
get_url(name = name, department = department)
get_url(name = name, university = university)
get_url(name = name, department = department, university = university)
}
