% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_info.R
\name{sentiment_info}
\alias{sentiment_info}
\title{Provides positive words and negative words extracted from comments, and tags from the website.}
\usage{
sentiment_info(u = url, y = 2018, word = "Positive")
}
\arguments{
\item{u}{A Character value indicating the url of the webpage corresponding to an instructor.}

\item{y}{A number indicating the user are interested in comments after that year.}

\item{word}{A string indicating the user are interested in positive words, or negative words, or tags.}
}
\description{
Provides positive words and negative words extracted from comments, and tags from the website.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
sentiment_info(u = url, y = 2018, word = "Negative")
}
