% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_plot.R
\name{sentiment_plot}
\alias{sentiment_plot}
\title{Sentiment Plot}
\usage{
sentiment_plot(url, y = 2018, word = "Positive")
}
\arguments{
\item{url}{A character value indicating the URL of the webpage corresponding to an instructor.}

\item{y}{A number indicating the user are interested in comments after that year.}

\item{word}{A string indicating the user is interested in positive words, negative words, or tags.}
}
\value{
A wordcloud
}
\description{
A wordcloud of sentiment information extracted from comments and tags from the website.
}
\examples{
url <- 'https://www.ratemyprofessors.com/ShowRatings.jsp?tid=2036448'
sentiment_plot(url = url, y = 2018, word = "Positive")
}
